# Speed up batch mapping