--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "cell-load"
-version = "0.7.13"
+version = "0.7.14"
 description = "Dataloaders for training models on huge single-cell datasets"
 readme = "README.md"
 authors = [