--- a/tests/test_mapping_strategy.py
+++ b/tests/test_mapping_strategy.py
@@ -3,7 +3,7 @@ import pytest
 import torch
 from pathlib import Path
 
-from cell_load.mapping_strategies import RandomMappingStrategy
+from cell_load.mapping_strategies import RandomMappingStrategy, BatchMappingStrategy
 from cell_load.dataset import PerturbationDataset
 from conftest import create_toml_config, make_datamodule
 
@@ -468,3 +468,110 @@ class TestRandomMappingStrategy:
                 len(mapping_strategy_uncached.split_control_pool["train"][cell_type])
                 > 0
             )
+
+
+class TestBatchMappingStrategy:
+    """Tests for BatchMappingStrategy functionality."""
+
+    @pytest.fixture(scope="class")
+    def dataset_with_batch_strategy(self, synthetic_data):
+        """Create a dataset configured to use the batch mapping strategy."""
+        root, cell_types = synthetic_data
+
+        config = {
+            "datasets": {"dataset1": "placeholder"},
+            "training": {"dataset1": "train"},
+        }
+        toml_path = create_toml_config(root, config)
+
+        dm = make_datamodule(
+            toml_path,
+            embed_key="X_hvg",
+            basal_mapping_strategy="batch",
+            cache_perturbation_control_pairs=False,
+            control_pert="P0",
+        )
+        dm.setup()
+
+        dataset = dm.train_datasets[0].dataset  # type: ignore
+        return dataset, dm, toml_path
+
+    @pytest.fixture(scope="class")
+    def indices_data_batch(self, dataset_with_batch_strategy):
+        """Extract control and perturbed indices for the batch mapping dataset."""
+        dataset, dm, toml_path = dataset_with_batch_strategy
+        cache = dataset.metadata_cache
+        control_mask = cache.control_mask
+        control_indices = np.where(control_mask)[0]
+        perturbed_indices = np.where(~control_mask)[0]
+
+        return {
+            "dataset": dataset,
+            "dm": dm,
+            "toml_path": toml_path,
+            "cache": cache,
+            "control_indices": control_indices,
+            "perturbed_indices": perturbed_indices,
+        }
+
+    def test_register_split_indices_groups_by_batch_and_cell_type(
+        self, indices_data_batch
+    ):
+        """Ensure control pools are keyed by (batch, cell_type)."""
+        dataset = indices_data_batch["dataset"]
+        control_indices = indices_data_batch["control_indices"]
+        perturbed_indices = indices_data_batch["perturbed_indices"]
+
+        strategy = dataset.mapping_strategy
+        assert isinstance(strategy, BatchMappingStrategy)
+
+        strategy.register_split_indices(dataset, "train", perturbed_indices, control_indices)
+
+        # Keys should be tuples of (batch, cell_type)
+        for key, pool in strategy.split_control_maps["train"].items():
+            assert isinstance(key, tuple) and len(key) == 2
+            batch_key, cell_type_key = key
+            # Check pool membership aligns with the key
+            for idx in pool:
+                assert dataset.get_batch(idx) == batch_key
+                assert dataset.get_cell_type(idx) == cell_type_key
+
+    def test_get_control_indices_same_batch_and_cell_type(self, indices_data_batch):
+        """Selected controls should share both batch and cell type with the perturbed cell."""
+        dataset = indices_data_batch["dataset"]
+        control_indices = indices_data_batch["control_indices"]
+        perturbed_indices = indices_data_batch["perturbed_indices"]
+
+        strategy = dataset.mapping_strategy
+        assert isinstance(strategy, BatchMappingStrategy)
+
+        strategy.register_split_indices(dataset, "train", perturbed_indices, control_indices)
+
+        pert_idx = int(perturbed_indices[0])
+        pert_batch = dataset.get_batch(pert_idx)
+        pert_cell_type = dataset.get_cell_type(pert_idx)
+
+        ctrl_idxs = strategy.get_control_indices(dataset, "train", pert_idx)
+        assert isinstance(ctrl_idxs, np.ndarray)
+        assert len(ctrl_idxs) == strategy.n_basal_samples
+
+        for ci in ctrl_idxs:
+            assert dataset.get_batch(int(ci)) == pert_batch
+            assert dataset.get_cell_type(int(ci)) == pert_cell_type
+
+    def test_random_state_reproducibility_batch(self, indices_data_batch):
+        """Same seed should yield identical control selections."""
+        dataset = indices_data_batch["dataset"]
+        control_indices = indices_data_batch["control_indices"]
+        perturbed_indices = indices_data_batch["perturbed_indices"]
+
+        s1 = BatchMappingStrategy(random_state=7, n_basal_samples=3)
+        s2 = BatchMappingStrategy(random_state=7, n_basal_samples=3)
+
+        s1.register_split_indices(dataset, "train", perturbed_indices, control_indices)
+        s2.register_split_indices(dataset, "train", perturbed_indices, control_indices)
+
+        pert_idx = int(perturbed_indices[0])
+        out1 = s1.get_control_indices(dataset, "train", pert_idx)
+        out2 = s2.get_control_indices(dataset, "train", pert_idx)
+        assert np.array_equal(out1, out2)