This PR speeds up dataloading when `basal_mapping_strategy=batch`.

1. It speeds up the dataloader initialization from about 70s to 6s. This is accomplished through making sentence creation in `_process_subset` more efficient. We were previously create long masks for each sentence (there were millions of sentences when grouped by batch); this could be sped up by some upfront array sorting.
2. It speeds up dataloader iteration from 21s to 5s (on 25 batches). In investigating this slowness, I discovered that repeated, non-local accesses to the h5 file caused slowness; the solution was to add an LRU cache on `fetch_gene_expression`.